+++ packages/tokens/src/colors.ts
@@ -1,20 +1,41 @@
 export const colors = {
-  white: '#FFF',
-  black: '#000',
+  white: "#FFF",
+  black: "#000",
+ 
+  benef100: "#00995D",
+  benef75: "#40B385",
+  benef50: "#80CCAE",
+  benef25: "#BFE5D7",
 
-  gray100: '#E1E1E6',
-  gray200: '#A9A9B2',
-  gray400: '#7C7C8A',
-  gray500: '#505059',
-  gray600: '#323238',
-  gray700: '#29292E',
-  gray800: '#202024',
-  gray900: '#121214',
+  secundary100: "#B1D34B",
+  secundary75: "#C4DE78",
+  secundary50: "#D8E9A5",
+  secundary25: "#EBF4D2",
 
-  ignite300: '#00B37E',
-  ignite500: '#00875F',
-  ignite700: '#015F43',
-  ignite900: '#00291D',
+  tertiary100: "#FFF0C7",
+  tertiary75: "#FFF4D5",
+  tertiary50: "#FFF7E3",
+  tertiary25: "#FFFBF1",
 
-  test: '#FFF',
-}
+  cooperado100: "#0A5F55",
+  cooperado75: "#478780",
+  cooperado50: "#84AFAA",
+  cooperado25: "#C2D7D5",
+
+  error100: "#ED1651",
+  error50: "#F1507C",
+
+  warning100: "#FFCB08",
+  warning75: "#FFD846",
+  warning50: "#FFE583",
+  warning25: "#FFF2C1",
+
+  gray100: "#E1E1E6",
+  gray200: "#A9A9B2",
+  gray400: "#7C7C8A",
+  gray500: "#505059",
+  gray600: "#323238",
+  gray700: "#29292E",
+  gray800: "#202024",
+  gray900: "#121214",
+};